changed type to better conversion

+++ src/parse_utils.c
@@ -6,7 +6,7 @@
 /*   By: altikka <altikka@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/27 12:21:02 by altikka           #+#    #+#             */
-/*   Updated: 2022/06/28 17:35:05 by altikka          ###   ########.fr       */
+/*   Updated: 2022/06/28 20:03:50 by altikka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ int	populate_grid(t_grid *grid, char *marks, int ofs)
 	return (1);
 }
 
-int	allocate_grid(t_grid *grid, int type)
+int	allocate_grid(t_grid *grid, size_t type)
 {
 	t_pos	*size;
 	int		i;
